Refactored code and file structure
