+++ client/src/components/Sections/TrainingSeries/CreateTSComponents/CreateTrainingSeries.js
@@ -5,7 +5,7 @@ import PropTypes from "prop-types";
 
 import { connect } from "react-redux";
 import { addTrainingSeries } from "store/actions/trainingSeriesActions";
-import ProgressCircle from "components/Misc/Progress/ProgressCircle";
+import ProgressCircle from "components/UI/Progress/ProgressCircle";
 
 //Styles
 import { withStyles } from "@material-ui/core/styles";
